Added test cases for kotlin dsl
