+++ src/test/resources/acceptance/assertion_tasks.gradle
@@ -1,0 +1,24 @@
+task existsTask {
+    doLast {
+        // -PtaskName=...
+        if (project.tasks.findByName(taskName) == null) {
+            throw new IllegalArgumentException("$taskName not found")
+        }
+    }
+}
+
+task showUploadDeployGateTaskNames() {
+    doLast {
+        println project.tasks.findByName("uploadDeployGate").dependsOn.collect {
+            if (it instanceof Task) {
+                it.name
+            } else if (it instanceof String) {
+                it
+            } else if (it.hasProperty("taskName")) {
+                it.taskName
+            } else {
+                it
+            }
+        }.flatten().join("\n")
+    }
+}
