+++ src/test/resources/project/settings.gradle
@@ -20,4 +20,4 @@ pluginManagement {
     }
 }
 
-rootProject.name = 'app'
+rootProject.name = "app"
